requirements.txt
python-dotenv==1.0.1
sqlalchemy==2.0.25
imagehash==4.3.1
Pillow==10.2.0
requests==2.31.0
PyYAML==6.0.1
openai==1.12.0
schedule==1.2.1
